+++ src/dev/canuk790/tilegame/entities/creatures/Player.java
@@ -4,7 +4,7 @@ import java.awt.image.BufferedImage;
 
 public class Player extends Creature{
 
-	public Player(float x, float y, BufferedImage texture) {
+	public Player(int x, int y, BufferedImage texture) {
 		super(x, y, texture);
 	}
